feat: Implement AI Service architecture with OpenAI integration
- Added AI service types and interfaces in `types.ts`.
- Created base AI service class in `base-service.ts`.
- Developed OpenAI service implementation in `openai.ts` for chat streaming.
- Introduced tool conversion utilities in `tool-converters.ts` for handling various AI service providers.
- Implemented message validation logic in `validators.ts`.
- Defined MCP types and schemas in `mcp-types.ts` for structured data handling.
- Added retry utility functions in `retry-utils.ts` for robust operation execution.
- Established Tauri MCP client for server communication and tool invocation.
- Updated chat message model to reference MCP types.
