+++ src/features/tools/ToolsModal.tsx
@@ -1,6 +1,7 @@
 import React, { useMemo } from 'react';
 import { useLocalTools } from '@/context/LocalToolContext';
 import { useMCPServer } from '@/hooks/use-mcp-server';
+import { MCPTool } from '@/lib/mcp-types';
 
 interface ToolsModalProps {
   isOpen: boolean;
@@ -10,7 +11,7 @@ interface ToolsModalProps {
 const ToolsModal: React.FC<ToolsModalProps> = ({ isOpen, onClose }) => {
   const { availableTools: mcpTools } = useMCPServer();
   const { availableTools: localTools } = useLocalTools();
-  const availableTools = useMemo(
+  const availableTools: MCPTool[] = useMemo(
     () => [...mcpTools, ...localTools],
     [localTools, mcpTools],
   );
