+++ src/lib/ai-service/index.ts
@@ -1,0 +1,28 @@
+// Re-export all types and interfaces
+export type { AIServiceConfig, IAIService } from './types';
+
+export { AIServiceProvider, AIServiceError } from './types';
+
+// Re-export validation utilities
+export { MessageValidator } from './validators';
+
+// Re-export tool conversion utilities
+export {
+  convertMCPToolsToProviderTools,
+  convertMCPToolsToCerebrasTools,
+} from './tool-converters';
+
+// Re-export base service class
+export { BaseAIService } from './base-service';
+
+// Re-export all service implementations
+export { EmptyAIService } from './empty';
+export { GroqService } from './groq';
+export { OpenAIService } from './openai';
+export { FireworksService } from './fireworks';
+export { AnthropicService } from './anthropic';
+export { GeminiService } from './gemini';
+export { CerebrasService } from './cerebras';
+
+// Re-export factory
+export { AIServiceFactory } from './factory';
